changed the logic of notifications with the help of an intermediate service to move the sending of messages to telegram outside of the transaction boundary
