+++ src/main/java/stepanova/yana/config/TelegramConfig.java
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Configuration;
 import org.telegram.telegrambots.meta.TelegramBotsApi;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
-import stepanova.yana.telegram.TelegramNotificationService;
+import stepanova.yana.notify.TelegramNotificationService;
 
 @RequiredArgsConstructor
 @Configuration
