+++ src/main/java/stepanova/yana/notify/NotificationService.java
@@ -1,4 +1,4 @@
-package stepanova.yana.telegram;
+package stepanova.yana.notify;
 
 public interface NotificationService {
     void sendMessage(String message);
